Force TCG when running virt-customize
Running virt-customize stucks on some environment and
is happening due to nested kvm(see [1]), until root cause
is identified and fixed, force TCG for running
virt-customize.

Also use the ANSIBLE_CONFIG if it's present, this will
help in getting same ansible output as we are getting
for other playbook run's in the job.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1565179

Change-Id: Ifa9effe0c4a1d43dca4d723017377d9d0387b0e6