GateStatus for hubbot to handle multiple changes
After this change the GateStatus plugin will be able to query and report
on multiple changes, allowing us to see reports from stable branches as
well.

Also add thorough testing for each function that can be tested with
`supybot-test --plugins-dir=".." GateStatus` when in the GateStatus
folder.

Also fetch Limnoria from pypi instead of using the source.

Change-Id: I73218c4d47a486f6e108813b9dd96cd741b95e57