promoter: rewrite hash exclusion critera
We are currently excluding some hashes from promotion based on timstamp
of the latest promoted hash. This proved to be unreliable in the current
asynchronous promotion process, where it's perfectly normal that a older
hash gets promoted after the next candidate is promoted.
This change rewrites the function taht implements the exclusion
criteria, now we are passing all the candidate hashes younger than the
latest promoted. The comparison is strictly maintained to the candidate
hashes, and the latest promoted hashes is not involved.

Change-Id: Id98d85dcedd36f02d00b5821f505d9a5cb454ab1
Closes-Bug: #1768090