Update patch set 4
Patch Set 4:

You're a member of the team. You know the board and the bug platform. This patch is out of scope for the sprint the team is in, it has no card associated, no related bug, I see no discussion, no context, no shared design, no priority.
A review for this would probably need at least 2 hours to be done properly (getting in context, understanding the design and the definition of done, testing iti, iterate on it) and I don't understand with which criteria I or any other core should allocate these hours of sprint time without a priority and all the information above. If any core is willing to spend the hours requested on this (please spend the proper amount of time to review this and agree to its design) please contact the UA to properly reduce the sprint workload.
Alternatively, contact the UA to associate this patch with a card to be properly inserted in a sprint, so it can follow the team process, with all that it means (discussion, shared design, ...). Or create a bug that needs to be properly triaged and prioritized.

Patch-set: 4
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0