Update patch set 1
Patch Set 1:

lgtm, have not run it yet (will this afternoon).  

I think we need this asap as we just failed last night on a missing container.

Question (not a -1, and more discussion for future iterations of this entire playbook)  Have we already looked at the roles in ansible?

again I think this patch is the right approach for today.

Patch-set: 1
Label: Code-Review=0