Update patch set 62
Patch Set 62: Code-Review-1

(7 comments)

-1 for grafyaml, I do think we should be reusing that here then writing something from scratch.

As for other comments, is there any sort of spec for this?  While some of this is rdo specific, others bits do apply to upstream openstack-infra.  Today we do collect a fair bit of info in graphite.o.o, grafana.o.o, openstack-health and other databases.  If there is something specific the tripleo project wants to be doing, we should be looking to work with the openstack-infra team to help reduce the amount of things that need to be maintained downstream.

There is also a line where, if this is needed for tripleo project, then we should be hosting this upstream in openstack, not rdoproject.

Patch-set: 62
Label: Code-Review=-1