promoter: Use force parameter when removing images
http://docs.ansible.com/ansible/latest/modules/docker_image_module.html

The current promoter does not actually untag and delete the
images from the promoter's local disk, yielding eventual
out of space issues for /var/lib/docker.

This adds the force option, which when used together with
state: absent should actually remove the bits.

(from doc: state)

"When absent an image will be removed. Use the force option to
un-tag and remove all images matching the provided name."

(from doc: force)

"Use with state absent to un-tag and remove all images matching the
specified name. Use with state present to build, load or pull an
image when the image already exists."

Change-Id: Ie0d801d167986696f72207d4ece571529c50daa2