Update patch set 1
Patch Set 1:

> I need more info about this.

Both the hashes mentioned in the patch are promoted to current-tripleo
https://trunk.rdoproject.org/centos7-master/current-tripleo/delorean.repo
https://trunk.rdoproject.org/centos7-queens/current-tripleo/delorean.repo

Without this patch promotion jobs are running with above hashes which is wrong, we need to run with the current consistent.

http://rhos-release.virt.bos.redhat.com:3030/rhosp

Also currently master RDO Phase 1 require a new promotion of consistent repo to unblock it so we need to merge this patch.

Patch-set: 1
Reviewer: Gerrit User 403 <403@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0