Make Ansible console output more readable for humans
Adds ANSIBLE_STDOUT_CALLBACK=debug which makes Ansible print stdout
and stderr results in an easy to read manner instead of JSON format
with collapsed newlines.

Do not be confused by "debug" in the callback name with debuging or
more verbosity, this was just a misfortunate naming mistake when
the human_log callback was merged into ansible core.

Change-Id: I4f7940ccfa0ddc53c61d4b83ff0762fd66eabd78
Partial-Bug: #1787912