Better PATH variable for Jenkins ooq jobs
Currently default PATH is taken from existing jenkins proccess
which is restriscted as comes from cron job. This patch makes
sure a complete PATH is used before starting oooq execution
which needs to use commands as 'ip' which is in /usr/sbin not
included by default.

Also this patch allows us to use tools install at user level
by adding ~/.local/bin to the PATH. This is the standard location
and also added to default path on newer operating systems like
Fedora 29. There are plans to backport similar change to other systems.
Still, even when implemented it would be impossible to affect user
accounts created after the day this change went effective.

What are the CI benefits? With this in place we can install a tool
and use it without having to alter the operating system. For example
we can use newer pip/tox/virtualenv/... version without the need to
sudo to upgrade it.

Change-Id: I02ad91a7dbc4eff98517fcbed5a8d2fa03c54ecf
See: https://fedoraproject.org/wiki/Changes/UserPathPrioritization