Fix selinux for te-broker log rotation
logrotation is still failing due to:
te-broker.rdocloud logrotate[16598]: ALERT exited abnormally with [1]
which indicates a selinux failure.

SELinux denies logrotate to check the attributes of the rotated log file,
this happens when logrotate has to rotate files outside of /var/log.

Observed that after manually setting selinux and rotating logs, selinux
property was not preserved, so it was missing from
/etc/selinux/targeted/contexts/files/file_contexts.local.

Now, selinux has been set and logrotate works with a -f (force):

/var/www/html/tebroker    system_u:object_r:var_log_t:s0
/var/www/html/tebroker/testenv-worker.log    system_u:object_r:var_log_t:s0

logrotate -v /etc/logrotate.d/te-broker

Enforcing selinux on prerotate and postrotate scripts so when the new log
file is created var_log_t is set.

Change-Id: I734e5b5d725cf258329f982bb3c83ea773c10376