Update patch set 4
Patch Set 4:

Feel free to update the message as I am not sure how to rephrase it to make it more clear but here is an example:

Assuming we install a newer version of tox/virtualenv/pip at user level (pip install --user ...), this is what happens upstream with ensure-tox role.

Installing at (ci) user level is a safe approach for enabling newer version of build tools without risking to mess with the operating system default ones.

The problem that you may face here is:

A) jobs running would fail to use the new tool becase old one is in PATH unless they use python module calling method (python -m foo), which is not affected by this.

B) tool appears to broken if the wrong executable is used which would try to use the upgraded module.

Very easy to reproduce on centos with tox, just do a "pip install --user -U tox" and run "/usr/bin/tox" after this.

Python gives precedence to user-libraries over sytem-ones, this means that tox executable will try to load the new tox module,... and guess what it may fail to import some bits because tox was internally refactored.

This can be avoided if ~/.local/bin would be in PATH, because the new executable will work.

Please read https://fedoraproject.org/wiki/Changes/UserPathPrioritization -- This should explain it better. It was approved by Fedora and there is bug for backporting the change to fedora-28. The Debian flavours already made this change back in 2016.

Because the fix is inside user account template it means that the only way to fix it on already deployed systems is to find a way to alter the user PATH.

Patch-set: 4
Reviewer: Gerrit User 371 <371@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0