Add ci reproducer using zuul quickstart
We need a way to reproduce CI jobs at zuul, but we don't want to emulate
zuul using bash scripts or the like, so we are testing here the
possibility of directly the samallest set of services needed to run
jobs.

For this first approach we have chose docker-compose with zuul provided
docker images to reduce the footprint of the user.

Story: https://tree.taiga.io/project/tripleo-ci-board/task/451?kanban-status=1447278

Change-Id: Ia36208599176ddfc91290e401ea7c0b60f979d5e