Change tripleo-upstream job to run once every four hours
This patch alters the trigger for the tripleo-upstream job to run
once very four hours from 1PM to 1AM UTC. This is due to the
current schedule of roughly once per day making it difficult to
test patches once landed. This change will be reverted once the
job is stable.

Change-Id: I85ea444a13df690fbe75ed0257fe1afac9d1dcc3