WIP: Separate the Centos-7 and Fedora-28 promotion ini files
This review adds more descriptive job names to the promotion
criteria in place of the job_id - which would have resulted in
duplicate jobs being considered in the case of a mixed-distro
pipeline.

This review adds new files but does not delete the old
ini files to allow for a staged adoption of the new
files.

Change-Id: I59d1deaa4e8d8b12cd1f561da2bbfa1280e89c00