WIP: Separate the Centos-7 and Fedora-28 promotion ini files
This review adds mpre descriptive job names to the promotion
criteria in place of the job_id - which would have resulted in
duplicate jobs being considered in the case of a mixed-distro
pipeline.

This review changes the master.ini file as an example.
If this new naming convention for RDO Cloud jobs is accepted,
all ini files will need to be updated.

Change-Id: I59d1deaa4e8d8b12cd1f561da2bbfa1280e89c00