Adjust ppc64le Triple O Build Job to Run Every 4 Hours
This patch is intended to work around a jenkins limitation that
prevents time definitions from wrapping around midnight. Due to this,
it is impossible to run the job in a time span that reaches beyond
4PM EST with the RDO default timezone. The easiest fix is to run
every 4 hours until the job is stable to test patches, then
revery to every 24 afterward.

See https://issues.jenkins-ci.org/browse/JENKINS-18313 for more
information.

Change-Id: I28c89c2c34622240257160b40d7572f1ca80b6dc