Separate jjb run from linters to jjb tox environment
Creates tox-jjb job that runs only jjb so we avoid running it inside
linters job which is bit confusing for users as jjb does not really do
that.

This will also allow us to remove tox-pep8 by running it as part of
linters (done in follow-up change).

Also adds the temporary token to gitignore, as this file could
remain on disk if you stop tox during execution.

Change-Id: Icc24b18ae476fd99df5681062a26f63a34f8d794