Update patch set 1
Patch Set 1:

Adding a bit of context here while waiting for the commit message. The current storage driver for docker is devicemapper, which has poor performance and has known issues with RH based distribution regarding the storage space. The space in the partition for docker is filled even when there's no image in the registry.
The performance is one of the reasons why we cannot promote two releases at the same time, the filesystem just explodes while deealing with containers.
Overlay2 seems to be the recommended choice for any version of docker, and requires no additional tools, just a working filesystem. We are testing this a bit with the server and see if things improve

Patch-set: 1