Refactor dlrn-promoter into a tiny service
dlrn-promoter service now is tiny script that runs in loop:
fetch latest config, install, call promoters, and sleep. Wake up,
repeat.

The advantage is that we can change how we call promoters without
having to reload the service. If we change the dlrn-promoter.sh
script, no changes are required to the service. For example, if
you want to add a new release or distro, update dlrn-promoter.sh
script only. The new script will be automatically updated and used
by the dlrn-promoter service.

Change-Id: I595385f2b3d8abb87042b2d8005dd02ddcdc0afd