Update tripleo-upstream-containers-build job to use tripleo-ci playbooks
We currently use some scripts that have been deprecated to build ppc64le
containers. This patch migrates the job to use the tripleo-ci playbooks
used by upstream. This requires some mocking of zuul values as we are
running in a jenkins environment. This has also allowed us to switch
our container image build backend to buildah.

Change-Id: I01c9d330e1bcc494badfd97e1d134f8c1acc2eda
Co-Authored-By: Adam C Kimball <baha@astyr.net>