promoter: no promotion if no job in criteria
Currently promoter checks only if successful jobs are missing from the criteria
so even if no jobs in criteria are specified, no missing jobs means we
can promote.
This patch adds a check so if there are no jobs in criteria, we are not
going to promote an untested hash.

Change-Id: I2bdca0b9001096c2739de170cde2a48bf095aff9
Related-Bug: #1823136