[DNM] Replace invalid groups with correct ones
Apparently this did not affect people already members of config-core becase gerrit did fallback to default ACL but
prevented normal users from even browsing the ACLs and see
who is core.

Zuul should not give a failure to merge on this change....

PS. This change is no longer needed, is still open only to see why
Zuul gives the weid failure to merge error on it.

Change-Id: I6d2d6dac805a85f9b22ba9e762d690d007a8403b