Remove gate-check part from hubbot
As last week hubbot gate-recheck usage was banned due to extra load
put on gerrit servers, we need to remove this functionality.

Adds test for execution of hubbot role via molecule so we know that
our change does not break the world, and avoid future regressions.

Cointains few minor fixes made to hubbot role that make it run
in our test environment.

Change-Id: I0c341367057af2c38cc1fd2cb53c426a95e21095