Improve ppc64le TripleO container build logs
This patch seeks to improve logging for the ppc64le tripleo container
builds by outputting the URL that they can be found at to the console,
so that someone looking at the job can easily navigate to the logs.
It also attempts to grab console text from the ci.centos job and
upload it alongside the container build jobs.

Change-Id: I472ba8cde8b192c14768161123b984517648d85c