Update patch set 22
Patch Set 22:

This is not so much about this role by itself but more about assuring that we can run molecule on rdo.

Regarding the role, its name is there for historical reasons but the functionality is still used.

What matters the most on this review is to see that molecule is called and runs.

Patch-set: 22
Label: Code-Review=0