Fixes pytest collection with missing imports
PyTest testes should be written in such way that running
`pytest --collect-only` would not fail when some dependencies are not
installed.

This is important because users may want to run limited set of tests
which without any package dependencies.

Without using this approach we prevent them from running pytest as
pytest would choke before even finishing the discovery.

Also fixed bug in py27/py36 triggering patterns which missed to run
on that change.

Change-Id: Idf085b1f29d656ce0cefe05db79485b0387ace91