Update patch set 36
Patch Set 36:

Wes, have you tried to see where te-broker role is imported?

Check rdo-infra/ci-config/ci-scripts/infra-setup/servers_setup.yml

Mainly "te-broker" role name is no longer about the old-te-broker, is about setting up the machine named te-broker, where we alsorun the hubbot.

Yes is it suspposed to be renamed to something else but nobody made a change to do that. I did not attempt the rename as I would not be able to test it without running the code.

Clearly rename is less important than testing.

Another aspect: the most important part this change was that it was removing the outdated molecule implementation with the standardized one, and allows us to easily add testing around other roles we have on ci-scripts.

Patch-set: 36
Label: Code-Review=0