promotion staging zuul job
Job definition and triggers for a staging promotion.

Configures a staging promoter using infra-setup role. This setup
will be used to run a staging promotion to test changes to the
promotion workflow and artifacts.

This change includes the basic provisioning for a staging promoter.
Tests and promoter configurations will be addressed in follow-up
reviews.

Story: https://tree.taiga.io/project/tripleo-ci-board/us/1105

Change-Id: I14601b5967ee2826fa9206c02325c45e23725ac4