promotion staging zuul job
Job definition and triggers for a staging promotion. Due to the nature
of the role (created to run in production promoter only), it is
expected to fail as it imports infra_setup/promoter role and run it as
is. A follow-up patch will handle environment conditions for running it
in nodepool node from a zuul job.

Story: https://tree.taiga.io/project/tripleo-ci-board/us/1105

Change-Id: I14601b5967ee2826fa9206c02325c45e23725ac4