Update patch set 11
Patch Set 11:

wes hayutin wrote:

> let's make sure we turn off the promoter service by default in the
> job as designed.
>
> 2019-08-21 22:28:53.254362 | TASK [promoter : Start and enable
> promoter] 2019-08-21 22:28:54.037564 | rdo-centos-7 | changed
>
> We can also turn off the monitoring w/ telegraph etc..

I disagree, this test is meant to check if any change to the promoter code will cause any disruption to the promotion process. What if a change in the monitoring part ends up breaking the workflow ? If we skip it, we don't get the necessary coverage.

Patch-set: 11
Label: Code-Review=0