WIP DNM create and push container manifest --arch ppc64le amd64
As tracked by tripleo-ci squad in [1] we must push a manifest
when pushing promoted containers. The manifest contains both
amd64 and ppc64le tagged containers. The parent patch in [2]
brings the required docker-ce and experimental mode needed
for manifest cli.

First creates manifest for containers built both in x86 and
ppc and then creates manifest for those that are x86 only.
The check for built ppc containers is done via url query
to [3].

[1] https://tree.taiga.io/project/tripleo-ci-board/us/1221
[2] https://review.rdoproject.org/r/#/c/21873/
[3] https://trunk.registry.rdoproject.org:8443/oapi/v1/namespaces/tripleomaster/imagestreamtags
Change-Id: I7a88e219d5a7c5a356b19ed3cef8c6d7890bb0ff