WIP DNM create and push container manifest --arch ppc64le amd64
As tracked by tripleo-ci squad in [1] we push a manifest
to dockerhub when pushing promoted containers. The manifest
contains both amd64 and ppc64le tagged containers.

For this initial iteration we are using the following for
the manifest name: {{container_name}}:{{full_hash}}_manifest
Eventually the name will be just name:hash but that will
follow in a subsequent patch.

For the manifest hash_x64_64 or hash_ppc64le containers
are pulled from rdo and pushed with those tags to docker.io.

Then create & push to docker.io a manifest with the arch tagged
containers you already pushed. The name will have _manifest for
now. If ppc is missing the manifest contains only x86.

Trying to add the include_tasks for testing in the promotion
staging job for now instead of molecule. WIP

[1] https://tree.taiga.io/project/tripleo-ci-board/us/1221
[2] https://review.rdoproject.org/r/#/c/21873/
Change-Id: I7a88e219d5a7c5a356b19ed3cef8c6d7890bb0ff
Depends-On: https://review.rdoproject.org/r/22464