Update patch set 33
Patch Set 33: Code-Review-1

(1 comment)

The -1 if for the duplicated task. I'm ok with the results and the general structure , but how do you check that the code here is correct if you'r not associating the tests ? The promotion stating job is passing, but there' no functional test that proves is passing in the right way.

Patch-set: 33
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1