Update patch set 7
Patch Set 7:

> @ykarel -  After a quick chat with Folco it seems like we should be
 > using tripleo-ci-testing rather than current-tripleo. I need to
 > figure out how to switch over to this though.
 > 
 > Overall, yes your understanding seems sound but I will reach out to
 > you to make sure we're on the same page.
 > 
 > I have added you as a watcher to the taiga.io board used for
 > planning this effort.

http://logs.rdoproject.org/openstack-periodic-master/opendev.org/openstack/tripleo-ci/master/periodic-tripleo-centos-7-master-containers-build-push/19076c5/job-output.txt.gz

tripleo-ci-testing is the hash to build containers and push to rdo registry.

when the promoter runs, it will push manifest and annotate for each arch, but that is already being implemented in https://review.rdoproject.org/r/22002.

Pushing current-tripleo wouldn't break anything but isn't the hash we want. I'm just +1 here for not having much context on what this job does. Logs results would be nice here too.

Patch-set: 7