Introduce multiple target registry and testing support for container-push
This change refactors container-push to support multiple target
registry.
We are going to push our images into quay.io as well and maybe one day
remove dockerhub, since we have reliability issues with it.
It also makes it possible for the container-push to be tested locally
The original playbook is not modified, a new playbook called
container-push-roled.yml is created instead, inactive in the production
workflow. When the new workflow will be ready, we can switch to it.

Change-Id: I9ff73c62817679f4a8225715b774a92434fa93ca