Update patch set 3
Patch Set 3:

(12 comments)

couple of comments on things that we could do as we're tweaking things. Also given we're looking to do multiple pushes to two different repos, we might want to look into changing the ansible strategy to "free", in that way we'd be able to execute all the tasks as we encounter them from each host (assuming there's more than one host). Another thing we might consider is breaking the playbook tasks up into a quay/docker/etc task file and just including it. If we do that then we'd have a clean template to follow should we ever need to extend or pivot to another registry.

Patch-set: 3
Label: Code-Review=0