Update patch set 4
Patch Set 4:

This will double the time of our promotion, and also double the possibility of having inconsistent containers hashes. There's a time during promotion in which the container images marked as "current-tripleo" in dockerhub do not point to the same hashes as the current-tripleo in DLRN. If soemthing fails during the upload in quay.io, the dlrn hsah will not be promoted, and we'll have the inconsistency for longer. Unfortunately promoting is not an atomic operation, there will always be an hash skew if we continue doing this.
I think we should try: 1) make the two uploads in parallel 2) use the docker API to retag the containers (the "promotion" retagging will be way faster)
When does this part need to be ready ?

Patch-set: 4
Label: Workflow=0