Simplifying the regex for prep image
This patch simplifies how to collect the prep image execution
time, since now we have this information at the end of the
task.
It also remove some return 0 for prep image in case it fails,
since we are using 0 as default if the key doesn't exist in
the build dictionary.

Change-Id: I6f6c90b1453189b9144e14d99d404e774c8a685b