Create and push container manifest --arch ppc64le amd64
As tracked by tripleo-ci squad in [1] we push a manifest
to the target registry (dockerhub, quay) when pushing
promoted containers. The manifest contains both amd64 and
ppc64le tagged containers. If ppc is missing the manifest
contains only x86.

For initial iteration we use the following manifest name:
{{container_name}}:{{full_hash}}_manifest. Eventually the name
will be just name:hash but that will be switched after testing.

To create manifests, hash_x86_64 and hash_ppc64le containers
are pulled from rdo and pushed with those tags to docker.io.
Then create & push the manifest with the arch tagged
containers you pushed.

Tested in the molecule-delegated ci job.

[1] https://tree.taiga.io/project/tripleo-ci-board/us/1221
Change-Id: I7a88e219d5a7c5a356b19ed3cef8c6d7890bb0ff