Promoter test: refactor staging env setup for tests
This patch refactor the staging environment script to include additional
informations in the stage info file, and be more specific in certain
informations. e.g we now know which commit is supposed to be promoted
and which one is the older promotion.
It also removes old unneeded files, and fixes some support in ansible
2.4

Change-Id: I6b531ddb623ad350ba06b78f8cbaa518e1488c29