Add promote-images molecule test and delegated-promote-images job
Adds a promote-images molecule scenario and job to run it which
sets up a staging promotion environment using
"--components overcloud-images". It then invoke promote-images.sh
promotion script and verifies the 'current-tripleo' and
'previous-current-tripleo' links are updated correctly.

This makes the promote-images.sh staging-server aware in that
there is now an explicit override if we are running staging by
detecting the /tmp/stage-info.yaml and using the images base
directory from there. Tracked by TripleO CI team in [1]

[1] https://tree.taiga.io/project/tripleo-ci-board/task/1314
Change-Id: I179d32f462a6664d9a4d1887d1e710d37c2d863a