Restore integration test
change https://review.rdoproject.org/r/22805 removed an essential part
of the new code that was breaking the current promotion in production.
This change restores the integration test by installing ansible via
package, hopefully in a compatible way with what the production server
has.
This should be the last patch we merge for the new code before we branch
this repo

Change-Id: I88bdb277458c016d519d0dab3afdb90845cf818d