Add serial test execution in pytest
Stagin env setup is not reentrant, so when tests are launched in
parallel, multople staging setups are run, and the conflict with each
other.
The previous proposed solution of adding another pytest command to run
only the serial tests is breaking the reporting, and it presents other
challenges in local testing actvivities.
This change leaves the mark serial for future reference, but disables
concurrency in tests.
Change-Id: I51d362e3a9cc5e3641ab2c1d4b3ea22ab5333434