Add serial test execution in pytest
Stagin env setup is not reentrant, so when tests are launched in
parallel, multople staging setups are run, and the conflict with each
other.
This change adds a new pytest command that disables parallelization on
the mark "serial".
Staging environment tests use the serial mark to spawn a single staging
setup
Change-Id: I51d362e3a9cc5e3641ab2c1d4b3ea22ab5333434