WIP! DNM remove _manifest suffix from container manifest name
Switches the default for manifest_push to true and removes
the _manifest suffix from pushed manifests. When we are pushing
manifests then we must not push containers with the same name
(name:hash). If manifest_push is true then we only push
container with named_label and not hash, before pushing manifests
with the hash.

Tracked by TripleO CI team at [1]. Posting WIP in preparation
but this must not merge until testing in [2] is complete.

[1] https://tree.taiga.io/project/tripleo-ci-board/task/1316
[2] https://tree.taiga.io/project/tripleo-ci-board/task/1315

Change-Id: I8a4c5762c8876ec93b11e7bdca2db03f721060c0