promotion staging: replace uWSGI with gunicorn
We are getting sporadic dlrnapi call failures, caused by uWSGI closing
connections, whose reason is very difficult to debug for the lack of
error messages.
It's probably related to load or response buffers, anyway, uWSGI would
need some fine tuning which is probably overkill for a app that serves
at most 5 requests and lives for 7 minutes top.
This change:
- Replace uwsgi with gunicorn. Seems to be fixing the issue (or it's just
  the multiple workers and even uWSGI would have worked)
- Change the port to the less overused 58080 to prevent any possible
  future conflict

Change-Id: I4bb4a52432c1a86517d45ae44baabe5b2657e73a