Update patch set 3
Patch Set 3: Code-Review+1

The earlier url was wrong(commit hash was not used at all), the correct should be using the commit id. As tripleo projects are not branched, it fallback to master, so using id is more correct as DLRN builds based on commit id, and same commit id should be looked in upstream repo for fetching container list.

Patch-set: 3
Label: Code-Review=+1