ppc64le: Set facts for delorean reporting to tripleo-ci-testing
Currently we do not do any delorean reporting for ppc64le containers.
We are also switching to building from tripleo-ci-testing rather than
from current-tripleo. This patch sets the ppc64le job up to report to
delorean for tripleo-ci-testing. It mimics how the delorean hash
information is generated in rdo jobs to achieve this. [0]

[0] https://github.com/rdo-infra/review.rdoproject.org-config/blob/master/playbooks/tripleo-ci-periodic-base/pre.yaml#L6-L14

Change-Id: Iedfebcb095db326aaf7e7fb8a088526099a913ff
Co-Authored-By: Adam C Kimball <baha@astyr.net>