ppc64le: Switch to tripleo-ci-testing for builds
We currently build containers based on current-tripleo,  and push
them to registry.rdoproject.org. We need to switch to
tripleo-ci-testing so that the manifest creation will be able to
find a pair of each containers based on the same hash. This patch
mimics how the delorean hash information is generated in rdo jobs
to achieve this. [0]

[0] https://github.com/rdo-infra/review.rdoproject.org-config/blob/master/playbooks/tripleo-ci-periodic-base/pre.yaml#L6-L14

Change-Id: Iedfebcb095db326aaf7e7fb8a088526099a913ff
Co-Authored-By: Adam C Kimball <baha@astyr.net>