Update patch set 1
Patch Set 1:

Hey Yatin,

The reason I think this is sufficient is because we are determining success with different criteria. The failure seems to be random and doesn't seem to prevent the container from building. I am more than happy to open a bug on tripleo about this though.

Sample can be seen here at 2019-10-25 15:06:31

https://centos.logs.rdoproject.org/tripleo-upstream-containers-build-master-ppc64le/1685/logs/logs/build.log

Do you think this is an error we should be concerned with?

Patch-set: 1
Reviewer: Gerrit User 514 <514@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0