Make sova-tracking-jobs non voting
This job was broken for more than a month and nobody attempted to fix
it. This is a good-enough reason for even removing a job from CI.

Making this non-voting until someone find time to either fix it or
remove it.

Change-Id: I2681a0e7e26aca1fdf79885df903258af349e73c
(cherry picked from commit f72ee4c71cf57353b8454b4e0edeb7b99de28f23)