Add support and test for secured registries
multi-arch manifest commands don't work well with local secured
self-signed registries, but we still need to test the roles with them,
especially to ensure that docker logins work.
This change adds support for secure registries for the staging
environment, using self-signed certificate.
But then it splits test into two categories, to get the coverage we
want:
- unsecure registries + multi-arch manifests push
- secure registries + login (multi-arch manifests are disabled)

Change-Id: Id697322cfa917b7da8c04c3f9d80ff8d079b1ed6