[WIP] Promotion rollback mechanism
The promotion process interacts with three different endpoints during a
promotion. None of the operations is atomic.
If a promotion fails in the middle of a run, the endpoints are left in a
inconsistent state. Containers may have a certain hash for a name, while qcow
images a different hash for the same name.

This change aims at creating a rollback mechanism to call in case of failure in
any of these stages.
The rollback tries to understand at what stage the promotion is, and tries to
return to a consistent status, minimizing the changes needed to do so (and the
possibility of something going wrong again)

Promoter Changes (trying to keep them minimal)
- The promoter workflow can now start a transaction at the start of a
  promotion, mark checkpoint for every stage completed, then end the transaction
- In case of failure, it can request a rollback on the created transaction

Components Added:
- a transaction object with rollback methods
- cli mode for the rollback operations.
- Basic class structure for promoter stages interaction
- pure python stages basic implementation, to maximize speed and minimize failures.

General Changes:

- moved staging-setup dir to staging_setup to allow python imports
- renamed all existing occurrences of staging-setup
- staging environment now able to handle multiple fixtures.
- staging environment able to handle dlrn server and repos
- staging environment injects a contaiers.yaml file
- zuul jobs now skip staging environment teardown, and collect
  promoter-staging subdir.
- staging environment puts every file generated under the same root dir
- staging environment create a full environment with qcow images
- staging environment make removal of local images pushed to source
  registry optional

Task: https://tree.taiga.io/project/tripleo-ci-board/task/1396
Change-Id: I5f8ac4fdeb732ed206fa6c323ed6db602832c833