Stabilize and improve promoter tests.
This changes touches different tests, to improve their stability or
readability

Changes in staging environment tests:

Staging env setup is not reentrant, so when tests are launched in
parallel, multiple conflicting staging setups are run.
The previous proposed solution of adding another pytest command to run
only the serial tests is breaking the reporting, and it presents other
challenges in local testing activities.
This change leaves the mark serial for future reference, but disables
concurrency in tests.

Additional changes in tox.ini:
- The method for testing in tox is slightly changed, parallelization is
  removed because staging environment is not reentrant
- rdo-tox-molecule now tests only non delegated tests
- added new molecule_delegated environment to isolate testing of
  delegated jobs

Additional changes in jobs configurations:
- container-push delegated job is now run under tox to get the nice
  reports.html
- rdo-tox-molecule is restricted to run on promoter role changes (still
  broken and not voting though)

Change-Id: I51d362e3a9cc5e3641ab2c1d4b3ea22ab5333434