ppc64le: Adjust Containers Job to Hold Cico Node
This patch adjusts the ppc64le container build job to set the cico
node held by it to a failed state, which prompts it to linger
after a build for 12 hours. It also adjusts the timing to happen
twice a day, during the work hours of myself and Mike Turek, so
that the job won't consume all cico nodes.

Change-Id: I7e51b673b15d7da9fcec750a64b381b56f854bcf