Correctly populate skiplist panel for all releases
Currently wrong data is populated for all releases
in skiplist panel, refer lp#1857334. This was due
to same dictionary('result') reference was added to
'parse_result' list, this is fixed by creating
new dictionary using deepcopy method and adding that
to the list instead of referencing same dictionary.

In addition to fixing the above issue this patch
also does following:-
- Remove unnecessary timestamp field from table load_pass_mysql.sql
- Remove unused imports
- Fix some indentation

Closes-Bug: lp#1857334
Change-Id: Ifa556d26ebcd8fd49a24e306951ce3b795f9c1fd
Signed-off-by: SurajP <surpatil@redhat.com>