ppc64le: Fix string output based on success or fail
We are outputting success or fail to stdout on the cico node.
This does not allow us to match text for the publisher as it
is not output on the Jenkins node's console.  This patch
moves the output to the Jenkins node allowing us to properly
match for the publisher.

Co-Authorer-By: Adam C Kimball <baha@astyr.net>
Change-Id: I43328849ed092e348b9c5364ef49f0c070c37caf