Make broken promoter tests work again
During the holidays, something changed and tests are not working
correctly anymore:

- Docker operations from zuul user started to fail with "Permission
denied". I suspect persistent connection were enable for ansible in
zuul, and a group add task was not taking effect like it was before.
Added a reset_connection after all group add operations. Added
functional test to ensure it stays checked (even if functional test are
not working right now)
- The hash the tests were using as candidate for container push does not
exist in dlrn server anymore, and the test are not isolated enough to
ignore that. Upstream test hit proxy cache but local test are failing,
and the cache will eventually expire. Change the hash in the fixture to
a more recent dlrn hash.

Change-Id: Idd9ae5271696b5c95601e3f115748909d10170a0