Promoter modularization: create initial transition
This change is the first step to prepare the code to be easily modified
for the component pipeline.
Most of the content is moved to a file to contain and maintain the
legacy code.
The main entry point is left on the original file and will initially
import and wrap functions from the legacy code, to create a favorable
environment in which replacing old functions with the new code or
reverting to the original workflow will be easier.

Change-Id: I0e2c5afa46a6ee1b3eace2cb761467aea85d08b0