Modularize Promoter: add config class
the legacy promoter was using single values to pass configuration
parameters to the functions, whose parameter list were growing to
contain even 11 arguments.
In the modularized design a single config object will be created and
passed to all the function/methods.
This change just add the config class and the test.
It is safe to merge as the object is not used anywhere in the code
and will be connected in a following patch.

Change-Id: I56c8550653311ad81933d30b2bcb8a06ec79f301